CC-1793. Addressed code-review comments.
